moved render_segmentation to pl module
