+++ spatial_data/pp/preprocessing.py
@@ -822,81 +822,6 @@ class PreprocessingAccessor:
         return obj.pp.add_observations()
 
 
-    def render_segmentation(
-        self,
-        alpha: float = 0,
-        alpha_boundary: float = 1,
-        mode: str = "inner",
-    ) -> xr.Dataset:
-        """
-        Render the segmentation layer of the data object.
-
-        This method renders the segmentation layer of the data object and returns an updated data object
-        with the rendered visualization. The rendered segmentation is represented in RGBA format.
-
-        Parameters
-        ----------
-        alpha : float, optional
-            The alpha value to control the transparency of the rendered segmentation. Default is 0.
-        alpha_boundary : float, optional
-            The alpha value for boundary pixels in the rendered segmentation. Default is 1.
-        mode : str, optional
-            The mode for rendering the segmentation: "inner" for internal region, "boundary" for boundary pixels.
-            Default is "inner".
-
-        Returns
-        -------
-        any
-            The updated data object with the rendered segmentation as a new plot layer.
-
-        Notes
-        -----
-        - The function extracts the segmentation layer and information about boundary pixels from the data object.
-        - It applies the specified alpha values and mode to render the segmentation.
-        - The rendered segmentation is represented in RGBA format and added as a new plot layer to the data object.
-        """
-        assert Layers.SEGMENTATION in self._obj, "Add Segmentation first."
-
-        color_dict = {1: "white"}
-        cmap = _get_listed_colormap(color_dict)
-        segmentation = self._obj[Layers.SEGMENTATION].values
-        segmentation = _remove_segmentation_mask_labels(segmentation, self._obj.coords[Dims.CELLS].values)
-        # mask = _label_segmentation_mask(segmentation, cells_dict)
-
-        if Layers.PLOT in self._obj:
-            attrs = self._obj[Layers.PLOT].attrs
-            rendered = _render_label(
-                segmentation,
-                cmap,
-                self._obj[Layers.PLOT].values,
-                alpha=alpha,
-                alpha_boundary=alpha_boundary,
-                mode=mode,
-            )
-            self._obj = self._obj.drop_vars(Layers.PLOT)
-        else:
-            attrs = {}
-            rendered = _render_label(
-                segmentation,
-                cmap,
-                alpha=alpha,
-                alpha_boundary=alpha_boundary,
-                mode=mode,
-            )
-
-        da = xr.DataArray(
-            rendered,
-            coords=[
-                self._obj.coords[Dims.Y],
-                self._obj.coords[Dims.X],
-                ["r", "g", "b", "a"],
-            ],
-            dims=[Dims.Y, Dims.X, Dims.RGBA],
-            name=Layers.PLOT,
-            attrs=attrs,
-        )
-        return xr.merge([self._obj, da])
-
     def render_label(
         self, alpha: float = 0, alpha_boundary: float = 1, mode: str = "inner", override_color: Union[str, None] = None
     ) -> xr.Dataset:
